server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.128.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "langchain[google]>=1.2.7",
    "langgraph>=1.0.7",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
